Stop storing cert & key by default for TLS
As pointed out in https://github.com/rhardih/serve/issues/3, storing the
generated certificate and key in the same directory as the one you're
serving with a file server, makes it possible, by default, to seize
them and do a MITM attack.

This change makes it so certificates are stored in memory by default
instead of on disk.

In some cases it might be nice to not have to accept an exception for a
new certificate in the browser, every time the server is started up and
for that two new flags have been added. --cert-save, --sc and
  --cert-dir, --cd.

These makes it possible to still store the certificate on disk, but
defaults to somewhere outside of the folder being served. (~/.serve)

This fixes #3.